Add comments and refactor
