+++ Sources/StreamVideo/WebRTC/v2/PeerConnection/Extensions/RTCRtpTransceiverDirection+CustomStringConvertible.swift
@@ -5,7 +5,20 @@
 import Foundation
 import StreamWebRTC
 
+/// Extends `RTCRtpTransceiverDirection` to conform to the `CustomStringConvertible` protocol.
 extension RTCRtpTransceiverDirection: CustomStringConvertible {
+    /// A textual representation of the RTP transceiver direction.
+    ///
+    /// This property provides a human-readable string for each possible direction of an RTCRtpTransceiver.
+    /// It's useful for debugging, logging, and displaying the current direction to users or developers.
+    ///
+    /// - Returns: A string describing the transceiver direction:
+    ///   - "sendRecv" for bidirectional communication
+    ///   - "sendOnly" for outbound-only communication
+    ///   - "recvOnly" for inbound-only communication
+    ///   - "inactive" when the transceiver is not actively sending or receiving
+    ///   - "stopped" when the transceiver has been stopped
+    ///   - "unknown/default" for any future, undefined directions
     public var description: String {
         switch self {
         case .sendRecv:
