+++ Sources/StreamVideo/WebRTC/v2/PeerConnection/Extensions/RTCVideoTrack+Sendable.swift
@@ -1,0 +1,12 @@
+//
+// Copyright © 2024 Stream.io Inc. All rights reserved.
+//
+
+import StreamWebRTC
+
+/// Extends `RTCVideoTrack` to conform to the `Sendable` protocol.
+///
+/// This extension uses the `@unchecked` attribute to indicate that the conformance
+/// to `Sendable` is not automatically checked by the compiler. The developer takes
+/// responsibility for ensuring thread safety when using this type across concurrency domains.
+extension RTCVideoTrack: @unchecked Sendable {}
