+++ Sources/StreamVideo/WebRTC/v2/PeerConnection/MediaAdapters/LocalMediaAdapters/LocalScreenShareMediaAdapter.swift
@@ -6,25 +6,53 @@ import Combine
 import Foundation
 import StreamWebRTC
 
+/// A class that adapts local screen sharing media for use in a streaming context.
+///
+/// This class manages the local screen sharing track, handling its publication,
+/// unpublication, and interaction with the peer connection.
 final class LocalScreenShareMediaAdapter: LocalMediaAdapting, @unchecked Sendable {
 
+    /// The unique identifier for the current session.
     private let sessionID: String
+    /// The peer connection used for WebRTC communication.
     private let peerConnection: StreamRTCPeerConnection
+    /// Factory for creating peer connection related objects.
     private let peerConnectionFactory: PeerConnectionFactory
+    /// Adapter for communicating with the Selective Forwarding Unit (SFU).
     private var sfuAdapter: SFUAdapter
+    /// Options for configuring video behavior.
     private let videoOptions: VideoOptions
+    /// Configuration settings for video.
     private let videoConfig: VideoConfig
+    /// Provider for screen sharing session information.
     private let screenShareSessionProvider: ScreenShareSessionProvider
 
+    /// The local video track used for screen sharing.
     private(set) var localTrack: RTCVideoTrack?
+    /// The type of screen sharing currently active.
     private var screenSharingType: ScreensharingType?
+    /// The video capturer used to capture screen content.
     private var capturer: VideoCapturing?
+    /// The RTP transceiver used to send the screen sharing track.
     private var sender: RTCRtpTransceiver?
 
+    /// The media stream identifier (mid) of the sender.
     var mid: String? { sender?.mid }
 
+    /// A subject for publishing track-related events.
     let subject: PassthroughSubject<TrackEvent, Never>
 
+    /// Initializes a new instance of the LocalScreenShareMediaAdapter.
+    ///
+    /// - Parameters:
+    ///   - sessionID: The unique identifier for the current session.
+    ///   - peerConnection: The peer connection used for WebRTC communication.
+    ///   - peerConnectionFactory: Factory for creating peer connection related objects.
+    ///   - sfuAdapter: Adapter for communicating with the Selective Forwarding Unit (SFU).
+    ///   - videoOptions: Options for configuring video behavior.
+    ///   - videoConfig: Configuration settings for video.
+    ///   - subject: A subject for publishing track-related events.
+    ///   - screenShareSessionProvider: Provider for screen sharing session information.
     init(
         sessionID: String,
         peerConnection: StreamRTCPeerConnection,
@@ -49,12 +77,18 @@ final class LocalScreenShareMediaAdapter: LocalMediaAdapting, @unchecked Sendabl
         screenSharingType = screenShareSessionProvider.activeSession?.screenSharingType
     }
 
+    /// Cleans up resources when the instance is being deallocated.
     deinit {
         sender?.sender.track = nil
     }
 
     // MARK: - LocalMediaManaging
 
+    /// Sets up the local media with the given settings and capabilities.
+    ///
+    /// - Parameters:
+    ///   - settings: The call settings to apply.
+    ///   - ownCapabilities: The capabilities of the local participant.
     func setUp(
         with settings: CallSettings,
         ownCapabilities: [OwnCapability]
@@ -62,6 +96,7 @@ final class LocalScreenShareMediaAdapter: LocalMediaAdapting, @unchecked Sendabl
         /* No-op */
     }
 
+    /// Publishes the local screen sharing track to the peer connection.
     func publish() {
         guard
             let localTrack,
@@ -96,6 +131,7 @@ final class LocalScreenShareMediaAdapter: LocalMediaAdapting, @unchecked Sendabl
         log.debug("Local screenShareTrack trackId:\(localTrack.trackId) is now published.")
     }
 
+    /// Unpublishes the local screen sharing track from the peer connection.
     func unpublish() {
         guard let sender, let localTrack else { return }
         Task {
@@ -117,6 +153,9 @@ final class LocalScreenShareMediaAdapter: LocalMediaAdapting, @unchecked Sendabl
         )
     }
 
+    /// Updates the call settings.
+    ///
+    /// - Parameter settings: The new call settings to apply.
     func didUpdateCallSettings(
         _ settings: CallSettings
     ) async throws {
@@ -125,6 +164,11 @@ final class LocalScreenShareMediaAdapter: LocalMediaAdapting, @unchecked Sendabl
 
     // MARK: - Screensharing
 
+    /// Begins screen sharing of the specified type.
+    ///
+    /// - Parameters:
+    ///   - type: The type of screen sharing to begin.
+    ///   - ownCapabilities: The capabilities of the local participant.
     func beginScreenSharing(
         of type: ScreensharingType,
         ownCapabilities: [OwnCapability]
@@ -161,6 +205,7 @@ final class LocalScreenShareMediaAdapter: LocalMediaAdapting, @unchecked Sendabl
         }
     }
 
+    /// Stops the current screen sharing session.
     func stopScreenSharing() async throws {
         try await sfuAdapter.updateTrackMuteState(
             .screenShare,
@@ -175,6 +220,9 @@ final class LocalScreenShareMediaAdapter: LocalMediaAdapting, @unchecked Sendabl
 
     // MARK: - Private helpers
 
+    /// Creates a new video track for screen sharing.
+    ///
+    /// - Parameter screenSharingType: The type of screen sharing to set up.
     private func makeVideoTrack(
         _ screenSharingType: ScreensharingType
     ) async throws {
@@ -219,5 +267,3 @@ final class LocalScreenShareMediaAdapter: LocalMediaAdapting, @unchecked Sendabl
         }
     }
 }
-
-extension RTCVideoTrack: @unchecked Sendable {}
