+++ Sources/StreamVideo/WebRTC/v2/PeerConnection/MediaAdapters/ScreenShareMediaAdapter.swift
@@ -105,13 +105,13 @@ final class ScreenShareMediaAdapter: MediaAdapting, @unchecked Sendable {
 
         // Set up observers for added and removed streams
         peerConnection
-            .publisher(eventType: RTCPeerConnection.AddedStreamEvent.self)
+            .publisher(eventType: StreamRTCPeerConnection.AddedStreamEvent.self)
             .filter { $0.stream.trackType == .screenshare }
             .sink { [weak self] in self?.add($0.stream) }
             .store(in: disposableBag)
 
         peerConnection
-            .publisher(eventType: RTCPeerConnection.RemovedStreamEvent.self)
+            .publisher(eventType: StreamRTCPeerConnection.RemovedStreamEvent.self)
             .filter { $0.stream.trackType == .screenshare }
             .sink { [weak self] in self?.remove($0.stream) }
             .store(in: disposableBag)
