+++ Sources/StreamVideo/WebRTC/v2/PeerConnection/MediaAdapters/VideoMediaAdapter.swift
@@ -102,13 +102,13 @@ final class VideoMediaAdapter: MediaAdapting, @unchecked Sendable {
 
         // Set up observers for added and removed streams
         peerConnection
-            .publisher(eventType: RTCPeerConnection.AddedStreamEvent.self)
+            .publisher(eventType: StreamRTCPeerConnection.AddedStreamEvent.self)
             .filter { $0.stream.trackType == .video }
             .sink { [weak self] in self?.add($0.stream) }
             .store(in: disposableBag)
 
         peerConnection
-            .publisher(eventType: RTCPeerConnection.RemovedStreamEvent.self)
+            .publisher(eventType: StreamRTCPeerConnection.RemovedStreamEvent.self)
             .filter { $0.stream.trackType == .video }
             .sink { [weak self] in self?.remove($0.stream) }
             .store(in: disposableBag)
