+++ Sources/StreamVideo/WebRTC/v2/PeerConnection/Models/PeerConnectionType.swift
@@ -4,7 +4,10 @@
 
 import Foundation
 
+/// Represents the type of peer connection in a WebRTC communication.
 enum PeerConnectionType: String {
+    /// Indicates a peer connection that is receiving media.
     case subscriber
+    /// Indicates a peer connection that is sending media.
     case publisher
 }
