+++ Sources/StreamVideo/WebRTC/v2/PeerConnection/Protocols/LocalMediaAdapting.swift
