+++ Sources/StreamVideo/WebRTC/v2/PeerConnection/Protocols/RTCPeerConnectionEvent.swift
@@ -1,0 +1,6 @@
+//
+// Copyright © 2024 Stream.io Inc. All rights reserved.
+//
+
+/// A simple protocol describing an event created from an RTCPeerConnection.
+protocol RTCPeerConnectionEvent {}
